requirements.txt
amqp==5.0.9
anytree==2.8.0
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black-widow==1.7.2
bs4==0.0.1
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cssselect==1.1.0
defusedxml==0.6.0
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.8
django-admin-thumbnails==0.2.6
django-allauth==0.42.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==5.9.0
django-cors-headers==3.10.0
django-countries==6.1.2
django-crispy-forms==1.13.0
django-currencies==0.10.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-oauth-toolkit==1.5.0
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
django-richtextfield==1.6.1
django-simple-history @ git+https://github.com/jazzband/django-simple-history@281d86928d4ccf520021b44354c99da1207029a9
django-sortedm2m==3.1.1
django-timezone-field==4.2.3
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
ExifRead==2.3.2
filelock==3.3.1
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
httpie==2.6.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
inflection==0.5.1
itemadapter==0.4.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonfield==2.1.1
jwcrypto==1.0
jwt==1.3.1
kombu==5.2.3
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
netifaces==0.11.0
numpy==1.21.1
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.5.3.56
packaging==21.3
parsel==1.6.0
phonenumbers==8.12.41
Pillow==8.3.2
pipreqs==0.4.11
platformdirs==2.4.0
priority==1.3.0
prompt-toolkit==3.0.27
Protego==0.1.16
proxyscrape==0.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyshark==0.4.3
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytidylib==0.3.2
pytz==2021.3
PyYAML==6.0
queuelib==1.6.2
redis==4.1.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
Scrapy==2.5.1
scrapy-djangoitem==1.1.1
scrapy-proxies==0.4
scrapy-proxy-pool==0.1.9
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
service-identity==21.1.0
simplejson==3.17.6
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlmap==1.5.12
sqlparse==0.4.2
stripe==2.64.0
termcolor==1.1.0
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing==3.7.4.3
typing-extensions==3.10.0.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
virtualenv==20.9.0
w3lib==1.22.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
yarg==0.1.9
zipp==3.6.0
zope.interface==5.4.0
